0.2.0 - The Experiments Begin
The NEWS.md file updated in this commit has much more detail, but in
short: I started experimenting with adding new features to the API.

What's here is the result of spending approximately 2 hours longer than
planned on messing around with the library. Check the commit metadata:
It's past 3am local time. I had to make myself stop trying stuff.

Not all of the extra functions and arguments are necessarily here to
stay, and nor is the 'instantlinks' search backend (which replaced
scraping the search result page HTML), but it was fun testing things.